feat: strip down all hard.fun code
