--- examples/hardDotFun/src/main.css
@@ -1,207 +1,0 @@
-@import 'bulma/css/bulma.css';
-
-:root {
-  --primary-color: #1a1a1a;
-  --accent-color: #f74928;
-  --accent-color-hover: #ff873e;
-  --text-color: #9e9c9b;
-  --text-color-hover: #bfbdba;
-  --text-color-focus: white;
-  --focus-border-color: #19b6e9;
-  --background-color: #121212;
-  --box-shadow: 0px 4px 10px rgba(0, 0, 0, 0.2);
-  --inner-shadow: inset 0 0 1px rgba(255, 255, 255, 0.1);
-  --border-radius: 10px;
-  --avatar-size: 100px;
-  --chat-button-size: 70px;
-}
-
-body {
-  height: 100vh;
-  margin: 0;
-  background: radial-gradient(
-      circle at -10% 70%,
-      #3c1a11ff 3%,
-      #3c1a1188 10%,
-      #1a0d0b66 25%,
-      #1a0d0b44 40%,
-      #1a0d0b22 60%,
-      #00000000 80%
-    ),
-    radial-gradient(
-      circle at 110% 20%,
-      #3c1a11ff 3%,
-      #3c1a1188 10%,
-      #1a0d0b66 25%,
-      #1a0d0b44 40%,
-      #1a0d0b22 60%,
-      #00000000 80%
-    );
-  background-size: cover;
-  background-repeat: no-repeat;
-  background-position:
-    -20% 70%,
-    120% 20%;
-  background-clip: padding-box;
-}
-
-.chat-button {
-  position: fixed;
-  bottom: 20px;
-  right: 20px;
-  width: var(--chat-button-size);
-  height: var(--chat-button-size);
-  border: 1px solid rgba(247, 73, 40, 0.75);
-  border-radius: 50%;
-  background-color: var(--primary-color);
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  cursor: pointer;
-  box-sizing: border-box;
-
-  img {
-    width: 100%;
-    height: 100%;
-    object-fit: contain;
-  }
-
-  &.open {
-    color: var(--accent-color);
-    background: var(--primary-color);
-    border: 1px solid var(--accent-color);
-  }
-}
-
-.iframe-root {
-  position: fixed;
-  bottom: 100px;
-  right: 5px;
-  width: 400px;
-  height: 700px;
-  background: var(--background-color);
-  border: 1px solid rgba(255, 255, 255, 0.08);
-  border-radius: var(--border-radius);
-  box-shadow:
-    0 4px 24px rgba(0, 0, 0, 0.4),
-    var(--inner-shadow);
-  display: none;
-  visibility: hidden;
-  overflow: hidden;
-
-  &.open {
-    z-index: 1000;
-    display: block;
-    visibility: visible;
-  }
-
-  iframe {
-    width: 100%;
-    height: 100%;
-    border: none;
-    border-radius: var(--border-radius);
-  }
-}
-
-.form-container {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: flex-start;
-  height: 100vh;
-  padding-top: 12.5%;
-
-  .field {
-    margin: 20px 0;
-  }
-
-  #form-header {
-    color: var(--text-color-focus);
-    font-size: 18px;
-    font-weight: 450;
-  }
-
-  .form-card {
-    max-width: 550px;
-    width: 100%;
-
-    input,
-    textarea {
-      border: 0.5px solid var(--text-color);
-      background: transparent !important;
-      border-radius: var(--border-radius);
-      font-size: 1.2rem;
-      color: var(--text-color);
-
-      &::placeholder {
-        color: var(--text-color);
-      }
-
-      &:hover {
-        border-color: var(--text-color-hover);
-      }
-
-      &:focus {
-        border-color: var(--focus-border-color);
-        color: var(--text-color-focus);
-        outline: none;
-      }
-    }
-
-    textarea {
-      max-height: 50px;
-    }
-
-    .button-connect {
-      width: 100%;
-      border-radius: 30px;
-      border: none;
-      font-weight: 800;
-      color: var(--text-color-focus);
-      background: linear-gradient(
-        45deg,
-        var(--accent-color),
-        var(--accent-color-hover)
-      );
-
-      &:hover {
-        background: linear-gradient(
-          45deg,
-          var(--accent-color-hover),
-          var(--accent-color)
-        );
-      }
-    }
-  }
-}
-
-.avatar-container {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  margin-bottom: 100px;
-}
-
-.avatar-circle {
-  width: var(--avatar-size);
-  height: var(--avatar-size);
-  border: 1px solid var(--text-color);
-  border-radius: 50%;
-  background-color: transparent;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  box-shadow: var(--box-shadow);
-  overflow: hidden;
-
-  img {
-    width: 100%;
-    height: 100%;
-    object-fit: cover;
-  }
-
-  svg {
-    width: 40px;
-    height: 40px;
-  }
-}
