--- examples/hardDotFun/src/window.d.ts
@@ -1,7 +1,0 @@
-import { MetaMaskInpageProvider } from '@metamask/providers';
-
-declare global {
-  interface Window {
-    ethereum: MetaMaskInpageProvider;
-  }
-}
