--- examples/hardDotFun/vite.config.ts
@@ -1,7 +1,0 @@
-import { defineConfig } from 'vite';
-import react from '@vitejs/plugin-react-swc';
-
-// https://vite.dev/config/
-export default defineConfig({
-  plugins: [react()],
-});
