+++ src/Conversation.tsx
@@ -2,12 +2,10 @@ import styles from './ChatView.module.css';
 import { Content } from './Content';
 import { StreamingText } from './StreamingText';
 import { messageStore, progressStore, toolCallStreamStore } from './store';
-import { ToolFunctions, GenerateCoinMetadataResponse } from './tools/types';
-import { TokenCard, TokenCardsStreamingPlaceholder } from './TokenCard';
 import type { ChatCompletionMessageParam } from 'openai/resources/index';
 import { memo } from 'react';
 import { useTheme } from './theme/useTheme';
-import { useAppContext } from './AppContext';
+import { useAppContext } from './context/useAppContext';
 
 export interface ConversationProps {
   messages: ChatCompletionMessageParam[];
@@ -50,46 +48,9 @@ const ConversationComponent = ({ messages, onRendered }: ConversationProps) => {
           );
         }
 
-        if (message.role === 'tool') {
-          const id = message.tool_call_id;
-          const prevMsg = messages[index - 1];
+        // if (message.role === 'tool') {
 
-          if (
-            !(
-              prevMsg.role === 'assistant' &&
-              prevMsg.content === null &&
-              Array.isArray(prevMsg.tool_calls)
-            )
-          ) {
-            return null;
-          }
-
-          const tc = prevMsg.tool_calls.find((tc) => tc.id === id);
-          if (!tc) return null;
-          if (tc.function.name !== ToolFunctions.GENERATE_COIN_METADATA) {
-            return null;
-          }
-
-          const toolResponse: GenerateCoinMetadataResponse = JSON.parse(
-            message.content as string,
-          );
-
-          return (
-            <div key={index} className={styles.left}>
-              <TokenCard
-                key={index}
-                id={toolResponse.id}
-                about={toolResponse.about}
-                symbol={toolResponse.symbol}
-                name={toolResponse.name}
-                prompt={
-                  JSON.parse(prevMsg.tool_calls[0].function.arguments).prompt
-                }
-                onRendered={onRendered}
-              />
-            </div>
-          );
-        }
+        // }
 
         return null;
       })}
@@ -123,11 +84,6 @@ const ConversationComponent = ({ messages, onRendered }: ConversationProps) => {
           </div>
         </div>
       )}
-
-      <TokenCardsStreamingPlaceholder
-        toolCallStreamStore={toolCallStreamStore}
-        onRendered={onRendered}
-      />
     </div>
   );
 };
