+++ src/stories/ChatView.stories.tsx
@@ -5,7 +5,7 @@ import { mockChatMessages } from '../mockdata';
 import { memeCoinGenIntroText } from '../config';
 import type { Decorator } from '@storybook/react';
 import { ThemeProvider } from '../theme/themeProvider';
-import { AppContextProvider } from '../AppContext';
+import { AppContextProvider } from '../context/AppContextProvider';
 
 const withProviders =
   (
